requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.2.6
open-python==1.1.1
openai==2.6.1
opencv-python==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.30.1
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
txtai==9.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
